Fix: Move custom_ui to all examples
